% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{analysegroups}
\alias{analysegroups}
\title{analysegroups identifies tentative groups of sites in a given block}
\usage{
analysegroups(blk, samdat, sitecol = "site")
}
\arguments{
\item{blk}{a statistical block or SAU containing sites of maturity samples}

\item{samdat}{a data.frame containing the raw maturity data used in the 
model fitting}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}
}
\value{
a list of nsites, siteproperties, site maturity parameters, the 
    fitted models, the tentative groups, and the selected data fro mteh input 
    data.frame
}
\description{
analysegroups uses a block and a data.frame of maturity data to
    conduct analyses that characterize the properties of all sites within a
    block that are present in the data.frame. It fits maturity ogives to each
    site, and identifies blocks based on attempts to combine sites in the
    sequential order of their latitudes, first low to high and then the 
    reverse.
}
\examples{
args(analysegroups)  #  ?analysegroups
print("wait on internal datasets")
# blk=13; samdat=samw; sitecol="site"
}
