% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{binglm}
\alias{binglm}
\title{binglm is a cleaner output from the summary of a binomial GLM}
\usage{
binglm(x, digits = 6, console = TRUE)
}
\arguments{
\item{x}{the binomial glm model}

\item{digits}{the number of digits printed for the coefficients}

\item{console}{should the result be printed to the console, default=TRUE}
}
\value{
the summary of the model (invisibly)
}
\description{
binglm should really be an S3 method for binomial GLM
    summary descriptions. I will make that conversion later. For now
    this outputs the information needed in a tighter form. It also
    outputs the summary (invisibly).
}
\examples{
\dontrun{
 # wait on a data set being included into the package
}
}
