% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{bootmature}
\alias{bootmature}
\title{bootmature generates bootstrap 95p CI for a maturity logistic curve}
\usage{
bootmature(
  samdat,
  label = "",
  lens = "length",
  mature = "mature",
  nb = 1000,
  lower = 50,
  upper = 160
)
}
\arguments{
\item{samdat}{the maturity dataset with a minimum of length and maturity
identified by the arguments 'lens' and 'mature'.}

\item{label}{a character string for a label to the model fit, default=''}

\item{lens}{the column name of the length data in the data.frame, default =
'length'}

\item{mature}{the column name of the maturity data 1 = mature, 0 = immature,
default='mature'}

\item{nb}{the number of bootstraps to perform, default=1000}

\item{lower}{the smallest size class for inclusion in the predicted plot 
data, the default = 50}

\item{upper}{the largest size class for inclusion in the predicted plot 
data, the default = 160}
}
\value{
a list of bootpar, bootout, and bootCI
}
\description{
bootmature takes the data used when fitting a maturity logistic
    curve, conducts 1000 bootstrap samples and refits the model to obtain 
    the parameters and predicted probabilities of maturity.
}
\examples{
print("wait on data sets")
# samdat=dat; label="";lens="length";mature="mature";nb=1000;lower=50;upper=160
}
\seealso{
{
  \link{fitmaturity}, \link{sample}
}
}
