% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{classicalL}
\alias{classicalL}
\title{classicalL the classical logistic curve}
\usage{
classicalL(p, lens)
}
\arguments{
\item{p}{a vector of three parameters, MaxDL, the maximum growth increment,
'a', the intercept of the exponential function if positive and b is
negative then we have an inverse logistic curve, and 'b' the gradient of 
the exponential function. If 'b' is positive and a is negative then a 
positive logistic curve is generated.}

\item{lens}{a vector of lengths for which the logistic value
will be calculated}
}
\value{
A vector of length(lens) containing the predicted logistic value at
    length values (can be used for growth increments, we use a statistical 
    method to estimate the same equation for maturity, with MaxDL = 1.0)
}
\description{
classicalL this uses the logistic function:
  exp(a+b*L)/(1+exp(a+b*L)), which has the property that the L50 = -a/b. We
  use this for growth increments and use a GLM with binomial errors to 
  obtain the same equation for maturity curves (taht assume maxDL=1.0).
}
\examples{
maxDL <- 5.0
a <- 6
b <- -0.1
lens <- seq(1,130,1)
growthinc <- classicalL(c(maxDL,a,b),lens)
}
