% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{comparetwo}
\alias{comparetwo}
\title{comparetwo compares two maturity curves fitted to two datasets}
\usage{
comparetwo(
  samdat1,
  samdat2,
  mature = "mature",
  length = "length",
  comparison = "season",
  lower = 50,
  upper = 160
)
}
\arguments{
\item{samdat1}{the selected maturity data for the first set}

\item{samdat2}{the selected maturity data for the second set}

\item{mature}{the name of the data.frame column containing the 1's and 0's
denoting mature and immature, default = 'mature'}

\item{length}{the data.frame column name of the length variable, defualt=
'length'}

\item{comparison}{the data.frame name of the comparison factor, default=
'season' the only current alternative is 'site'}

\item{lower}{the smallest size class for inclusion in the predicted plot 
data, the default = 50}

\item{upper}{the largest size class for inclusion in the predicted plot 
data, the default = 160}
}
\value{
a list of lists containing all models, the anova, and the probabilty  
   the curves are different.
}
\description{
comparetwo takes two datasets selected on a categorical factor
   then fits a maturity curve mature = length + comparison, and compares 
   that with a maturity curve mature = length. This requires the MASS library
   for the version of anaova that permits test='Chisq', which conducts an
   analysis of deviance (see Venables and Ripley, 2nd ed, p288). The function
   returns the expanded model, the combined model, and a maturity model for 
   each of the separate data sets.
}
\examples{
\dontrun{
  dat1 <- selecttwo(samdat=sam,value1=site,value2=c(seas1),
                    field1="site",field2="season")
  dat2 <- selecttwo(samdat=sam,value1=site,value2=c(seas2),#c("Autumn","Winter"),
                    field1="site",field2="season")
  ans <- comparetwo(dat1,dat2)
  str(ans,max.level=1)
}
}
