% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{comparetwogroups}
\alias{comparetwogroups}
\title{comparetwogroups compares two groups of sites using logistic regression}
\usage{
comparetwogroups(
  grp1,
  grp2,
  samdat,
  sitecol = "site",
  length = "length",
  mat10 = "mature",
  sigdiff = 0.05
)
}
\arguments{
\item{grp1}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{grp2}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}

\item{sigdiff}{what probability level constistutes a significant difference.
default=0.05}
}
\value{
a list of the anova between the models with and without group/site,
    the significance of any difference, model1, model2, and the outputs from
    applying binglm to each model
}
\description{
comparetwogroups using logistic regressions to compare whether 
    two groups of sites are sufficiently similar for their data to be 
    combined, or not.
}
\examples{
print("wait on internal data sets")
# grp1=g1;grp2=g2; sitecol="site";samdat=samb;
# length="length";mat10="mature"; sigdiff=0.05
}
\seealso{
{
   \link{getloc}, \link{siteprops}, \link{binglm}
}
}
