% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{comparetwosites}
\alias{comparetwosites}
\title{comparetwosites compares two groups of sites using logistic regression}
\usage{
comparetwosites(
  site1,
  site2,
  samdat,
  sitecol = "site",
  length = "length",
  mat10 = "mature",
  siglevel = 0.05
)
}
\arguments{
\item{site1}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{site2}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}

\item{siglevel}{what P level constitutes a significant difference, 
default = 0.05}
}
\value{
a list of site properties, and a summary of the logistic regression
}
\description{
comparetwosites using logistic regressions to compare whether 
    two groups of sites are sufficiently similar for their data to be 
    combined, or not.
}
\examples{
print("wait on internal data sets")
# site1=c(784,459,500);site2=c(499); sitecol="site";samdat=samb;
# length="length";mat10="mature";siglevel=0.05
}
\seealso{
{
   \link{getloc}, \link{siteprops}
}
}
