% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{doseR}
\alias{doseR}
\title{doseR the dose response estimate of deltaL}
\usage{
doseR(p, lens)
}
\arguments{
\item{p}{the parameter list, only the first three value are used here.
Typically p would also contain the sigma value representing the standard
deviation of the additive normal random residuals expected about the
fitted curve.}

\item{lens}{the initial or starting lengths}
}
\value{
a vector of the same length as nrow(x)
}
\description{
doseR calculates the dose response estimates of deltaL growth
    from input parameters a, b, and c: deltaL <- a/(1 + ((L/b)^c))
}
\examples{
p=c(1.46,11.0,6.5,0.25) # a, b, c, sigma, the latter not used here
# x=pdat
print("wait on internal data")
}
