% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{findgroups}
\alias{findgroups}
\title{findgroups sequentially compares sites and groups by maturity curve}
\usage{
findgroups(sitenums, samdat, sitecol = "site", console = FALSE)
}
\arguments{
\item{sitenums}{a vector of multiple site number from the data.frame samdat. 
They should be ordered according to Lm50, latitude or longitude, or some 
other factor that might suggest grouping}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{console}{should the group structure be echoed to the console as they 
are found? default=FALSE}
}
\value{
a list of groups of site numbers found to be statistically close
}
\description{
findgroups conducts a sequential comparison of logistic maturity
    model fits and combines site numbers with statistically similar curves.
    Ideally, to test for homogeneity of the groups found, whatever ordering 
    is used shoulkd be reversed and the findgroups applied again.
}
\examples{
print("wait on internal data sets")
}
