% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{fitDR}
\alias{fitDR}
\title{fitDR fits a dose response model to tagging data using Normal errors}
\usage{
fitDR(p, dat, Lt = "Lt", DL = "DL")
}
\arguments{
\item{p}{the parameters a, b, c, and sigma the sd for mormally distributed
residual errors}

\item{dat}{a data.frame containing at least Lt (initial length) and DL (delta
L)}

\item{Lt}{the name of the initial length data}

\item{DL}{the name of the increment in length data}
}
\value{
a list of results
}
\description{
fitDR fits a dose response, as defined by Rogers-Bennett for
    tagging data in Rogers-Bennett et al (2003). DeltaL <- a/(1 + ((L/b)^c)).
    Note well, this equation only works for annual data.
}
\examples{
# Add a sd function to the arguments, default = constant sd 
}
