% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{fitexp}
\alias{fitexp}
\title{fitexp fits an exponential model}
\usage{
fitexp(xdat, ydat)
}
\arguments{
\item{xdat}{the values for the independent variable, eg length}

\item{ydat}{the dependent variable, eg weight}
}
\value{
a list of the model, x and y, being the fitted curve, a and b 
    the fitted coefficients, and n the number of observations, and the raw 
    data
}
\description{
fitexp fits an exponential model, such as a weight-at-length
    model. Anything that uses a ydat = a.xdat^b form can be fitted by this 
    function.
}
\examples{
x <- seq(50,200,25)
y <- c(15,55,130,275,490,790,1220)
ans <- fitexp(x,y)
cat("a = ", ans$a, " and b = ",ans$b,"\n")
}
