% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{fitgroups}
\alias{fitgroups}
\title{fitgroups fits a maturity ~ length logistic regression to groups}
\usage{
fitgroups(
  groups,
  samdat,
  sitecol = "site",
  length = "length",
  mat10 = "mature"
)
}
\arguments{
\item{groups}{a list containing vectors of site numbers}

\item{samdat}{a data.frame of maturity data containing the required site(s)}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}
}
\value{
an invisible list containing the parameters and the fitmaturity 
    results
}
\description{
fitgroups after using findgroups one can use fitgroups to 
    fit maturity curves to each group found
}
\examples{
print("wait on internal data sets")
#  groups=groups1; samdat=samb;length="length";mat10="mature";sitecol="site"
}
