% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{fitgrow}
\alias{fitgrow}
\title{fitgrow fits a tagging growth model to tagging data}
\usage{
fitgrow(p, grow, sdfunc, dat, hessian = FALSE)
}
\arguments{
\item{p}{the parameters a, b, c, sigma, and tau}

\item{grow}{the function describing the growth function used}

\item{sdfunc}{the function describing the sd vs predicted DL}

\item{dat}{a data.frame containing at least Lt (initial length) and DL (delta
L) in columns 1 and 2}

\item{hessian}{should the hessian be generated? default=FALSE}
}
\value{
the fitted model output from the nlm function
}
\description{
fitgrow fits a selected tagging growth model to a set of 
    initial lengths and consequent growth increments. It requires the growth 
    model function to be input as a function argument as well as another 
    function, as a second argument, that is used to describe how the variance
    is expected to change with the changing predicted growth increments. The
    function uses negLLG to calculate the negative log-likelihoods
}
\examples{
print("waiting to be developed")     
}
