% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{fitmaturity}
\alias{fitmaturity}
\title{fitmaturity fits a maturity ~ length logistic regression to maturity}
\usage{
fitmaturity(
  sites,
  samdat,
  sitecol = "site",
  length = "length",
  mat10 = "mature"
)
}
\arguments{
\item{sites}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data containing the required site(s)}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}
}
\value{
an invisible list containing the regression results, the model,
    the parameters, and the plotdata
}
\description{
fitmaturity uses data of the form maturity = 1 immaturity = 0,
    and length and fits a glm using binomial errors (a logistic regression).
    It outputs the regression results, the parameters, a, b, Lm50, and IQ, as
    well as the data for plotting as 'lens', 'propm', and 'predm'
}
\examples{
print("wait on internal data sets")
#  sites=grp; samdat=samb;length="length";mat10="mature";sitecol="site"
}
