% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{fitsiteweights}
\alias{fitsiteweights}
\title{fitsiteweights fits each site's weight-at-length curve from an SAU}
\usage{
fitsiteweights(
  blk,
  dat,
  sau = "block",
  site = "site",
  len = "length",
  wt = "wholewt"
)
}
\arguments{
\item{blk}{the block or SAU containing at least one site}

\item{dat}{the dataframe containing at least site, length, and wholewt 
columns}

\item{sau}{name of the statistical blocks being used, default='block'}

\item{site}{name of the site numbers used, default='site'}

\item{len}{name of the length variable used, default='length'}

\item{wt}{name of the weight variable used, default='wholewt'}
}
\value{
a list of the model and x and y the predicted exponential curve
}
\description{
fitsiteweights fits an exponential curve to the data from each 
    site within an sau, along with an analysis of the sau data combined. The
    names of the variables used are defined below, they will be natural-log
    transofmred within fitexp so the raw values are required here.
}
\examples{
print("wait for data sets")
}
