% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{getdist}
\alias{getdist}
\title{getdist extracts the relative distances between sample sites}
\usage{
getdist(x, lat = "lat", lon = "long", site = "site")
}
\arguments{
\item{x}{the data.frame containing the maturity data}

\item{lat}{the name of the latitude variable}

\item{lon}{the name of the longitude variable}

\item{site}{the name of the site variable}
}
\value{
a square matrix containing the approximate distance between
    sites in kms
}
\description{
getdist expects maturation data from multiple sites 
    that includes latitude and longitude data for the location of
    each site. It calculates the approximate crow-flies distance 
    between the sites using the getlatlong and heversine functions, 
    and then gives the volumetric distance.
}
\examples{
\dontrun{
 # wait on a data set being included into the package
}
}
