% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{getdistances}
\alias{getdistances}
\title{getdistances estimates the haversine distance between sites}
\usage{
getdistances(siteprop)
}
\arguments{
\item{siteprop}{the output from 'siteprops' containing 'combsite', 'lat' and
'long' as a minimum}
}
\value{
a square symmetric matrix of haversine distances (kms) between all
    sites in a block
}
\description{
getdistance uses theoutput from siteprops to estimate the
    haversine distances between sites. It output a symmetrical matrix of
    these values.
}
\examples{
print("wait on internal data sets")
}
