% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{getlatlon}
\alias{getlatlon}
\title{getlatlon extracts the latitude and longitudes of sample sites}
\usage{
getlatlon(x, lat = "lat", lon = "long", site = "site")
}
\arguments{
\item{x}{the data.frame containing the maturity data}

\item{lat}{the name of the latitude variable}

\item{lon}{the name of the longitude variable}

\item{site}{the name of the site variable}
}
\value{
a matrix of sites by latitude longitude
}
\description{
getlatlon expects maturation data from multiple sites 
    that includes latitude and longitude data for the location of
    each site. It extracts and outputs the lat long for each site.
}
\examples{
\dontrun{
 # wait on a data set being included into the package
}
}
