% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{getloc}
\alias{getloc}
\title{getloc pulls out the lat longs and block for identified site numbers}
\usage{
getloc(sits, samdat, orderby = "none")
}
\arguments{
\item{sits}{a vector of site numbers to be found in the data.frame samdat}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{orderby}{this can be 'lat', 'long', or 'none'. If 'none' no reordering
will be done, otherwise the sites will be sorted by lat or long. default
= 'none'}
}
\value{
a matrix of site by lat, long, block, and sitename
}
\description{
getloc isa utility function that extracts the lat longs, if 
    present, for all the identified site numbers
}
\examples{
print("wait on internal data sets")
# sits=bsites; samdat=samb; orderby="none"
}
