% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{getsiteparams}
\alias{getsiteparams}
\title{getsiteparams get all the individual site parameters in a data.frame}
\usage{
getsiteparams(
  samdat,
  sitecol = "site",
  length = "length",
  mature = "mature",
  sau = "block",
  lower = 50,
  upper = 160
)
}
\arguments{
\item{samdat}{a data.frame of maturity data containing sites}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mature}{column name for the maturity variable in samdat mature=1,
immature = 0}

\item{sau}{column name for the sau or block in samdat, default='block'}

\item{lower}{the smallest size class for inclusion in the predicted plot 
data, the default = 50}

\item{upper}{the largest size class for inclusion in the predicted plot 
data, the default = 160}
}
\value{
a list containing a matrix of the site parameters and a list of all
    the model fits used (called allfit)
}
\description{
getsiteparams estimates the site parameters and all the model
    fits for all sites within a data.frame. This sets up the system ready
    for site comparisons
}
\examples{
args(getsiteparams)
print("wait on internal data sets")
samdat=samnw; sitecol="site";length="length";mature="mature";sau="block";lower=20;upper=160
}
