% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{groupprops}
\alias{groupprops}
\title{groupprops tabulates the properties of each site number}
\usage{
groupprops(site1, site2, samdat, sitecol = "site")
}
\arguments{
\item{site1}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{site2}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}
}
\value{
a list of the site properties, the distance between the two sites, 
    and a summary of the logistic regression
}
\description{
groupprops extracts the main properties of two vectors of site
    numbers. These properties include: n, site, sitename, lat, long, block,
    subblock, date, and propM (proportion mature). It uses siteprops
}
\examples{
print("wait on internal data sets")
# site1=as.character(c(780,783)); samdat=rdat; sitecol="site"
# site2=as.character(c(780,783)); 
}
\seealso{
{
   \link{getloc}, \link{siteprops}
}
}
