% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmaps.R
\name{haversine}
\alias{haversine}
\title{haversine calculates the distance between two lat-lon pairs}
\usage{
haversine(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{The first latitude in decimal degrees}

\item{lon1}{The first longitude in decimal degrees}

\item{lat2}{The second latitude in decimal degrees}

\item{lon2}{The second longitude in decimal degrees
#'}
}
\value{
a vector of the polar, volumetric, and equatorial distances as 
    kilometres.
}
\description{
haversine uses the Haversine formula to calculate the
    great circle distance between two pairs of lat-lon values. No
    projection is used so this is only useful for points close 
    together on the Earth's surface. In addition, it uses a NASA
    estimate of the volumetric radius of the Earth (6371). The radius
    estimate at the equator is 6378.137 and that of the polar radius
    is 6356.752. It produces a vector of all three of these values. The 
    differences tend to be minor so for biological samples the central 
    volumetric value should suffice.
}
\examples{
\dontrun{
  haversine(-43.31,145.76,-43.19,145.87) #16.09593, 16.132, 16.15007
}
}
