% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{ILfuns}
\alias{ILfuns}
\title{ILfuns prints syntax of the most used growth functions in biology}
\usage{
ILfuns()
}
\value{
nothing, but it does print some lines to the console
}
\description{
ILfuns prints the syntax of the most used growth functions in
    the package biology. This is designed to assist with using the syntax.
}
\examples{
ILfuns()
}
