% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{labelmake}
\alias{labelmake}
\title{labelmake combines the contents of a character vector to make a label}
\usage{
labelmake(vect, sep = "_", digits = 3)
}
\arguments{
\item{vect}{a vector of either all text or all numbers, mixtures not allowed}

\item{sep}{what separator should be used, default = '_'}

\item{digits}{if a numeric vector is input how many decimal places to be used?
default = 3}
}
\value{
a character string of the contents of vect combined
}
\description{
labelmake is a utility function that takes a vector of values, 
    either numeric or text, and combines them with a selected separator. Any
    numeric values also have the option of being rounded to a given precision
}
\examples{
x <- c("one","two","three")
labelmake(x)
x <- c(1,2,3,pi)
labelmake(x,digits=5)
}
