% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{negLLG}
\alias{negLLG}
\title{negLLG calculates the negative log-likelihood for growth models}
\usage{
negLLG(p, grow, sdfunc, x, pen = TRUE)
}
\arguments{
\item{p}{the vector of parameters, what these are will depend on both grow
and sdfunc}

\item{grow}{the function describing the growth function used}

\item{sdfunc}{the function describing the sd vs predicted DL}

\item{x}{a data.frame or matrix containing at least Lt (initial length) 
and DL (deltaL) in columns 1 and 2}

\item{pen}{a penalty to keep p[2] positive. default = TRUE}
}
\value{
a single number, the negative log-likelihood
}
\description{
negLLG is a general function for calculating the negative log-
    likelihood using additive normal random errors for tagging based growth 
    models defined in the functions 'grow' and 'sdfunc'. The 'grow' 
    argument must be a function that defines the expected growth increments 
    given two input arguments 'p' and 'Lt', where p is the vector of 
    parameters, and 'x' (or whatever) is a vector of initial lengths. The
    sdfunc is a function that defines the description of the standard 
    deviation around the predicted growth increments. A penalty is imposed on
    parameter[2], which is often L50 to avoid negative numbers
}
\examples{
 pars <- c(2.2,9.5,5,0.6,1)
 dat <- NULL  # wait on internal dataset
 # negLLG(p=pars,grow=invlog,sdfunc=sdpow)
}
