% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{outIL}
\alias{outIL}
\title{outIL provides a tidy summary of the output from fitIL}
\usage{
outIL(object, console = TRUE)
}
\arguments{
\item{object}{the object output from fitIL}

\item{console}{should the output be printed to the console, default=TRUE}
}
\value{
a character vector of 16 statements ready to be printed
}
\description{
outIL provides a tidy summary of the output from fitIL
}
\examples{
print("waiting to be developed")
}
