% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{plot.IL}
\alias{plot.IL}
\title{plot.IL an S3 method to plot an IL object}
\usage{
\method{plot}{IL}(
  x,
  outliers = FALSE,
  xmin = NA,
  xmax = 0,
  ymin = -3,
  ymax = 0,
  nbreaks = 25,
  ...
)
}
\arguments{
\item{x}{an IL object from the fitIL function}

\item{outliers}{identify outliers on the plot. Default = FALSE}

\item{xmin}{allows one to set the lower bound of the x-axis on growth plots.
default=NA, which means it will be taken from the data. On ecannot set a 
minimum greater than the minimum of the data}

\item{xmax}{allows one to set a particular maximum growth increment.
default=0 so that xmax is set by the data.}

\item{ymin}{allows one to set the lower bound of the growth plots. Default=
-3 as befits abalone}

\item{ymax}{allows the setting of the largest y-axis value. defualt=0, which
means it is set via the data}

\item{nbreaks}{number of bins in the histogram of data density, default=25}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
nothing but it does generate a plot
}
\description{
plot.IL is an S3 method for plotting the results of
    fitting an inverse logistic curve to a set of tagging data.
    it plot the data and the model fit, with curves depicted the
    90th percentile spread and, if outliers was set to TRUE in the
    call to fitIL, it identifies outliers as red dots, and removes
    those points from the calculation. It plots the residuals, the
    rate of change in DL, to find the turnover length, and the
    density of data by initial length
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
plot(ans,outliers=TRUE)
# x=ans; outliers=TRUE; xmin=NA; xmax=0;ymin=-2; ymax=0;nbreaks=25
}
