% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{plotcomparison}
\alias{plotcomparison}
\title{plotcomparison generates a plot from the output of comparetwo}
\usage{
plotcomparison(ans, name1, name2, P = 0.01)
}
\arguments{
\item{ans}{the output from the function 'comparetwo'}

\item{name1}{a character strong as the name of the first dataset}

\item{name2}{a character strong as the name of the second dataset}

\item{P}{the probabiolity lelve that is taken as representing a significant
difference between two maturity curves, default = 0.01}
}
\value{
nothing but it does plot a graph
}
\description{
plotcomparison uses the output of the function 'comparetwo' to
   generate a plot of the maturity curves fitted to the original data from 
   the two datasets, plus, if there is no significant difference, the 
   combined maturity curve, all with 95 percent CI.
}
\examples{
print("wait on internal data sets")
}
