% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{plotgroups}
\alias{plotgroups}
\title{plotgroups fits a maturity ~ length logistic regression to groups}
\usage{
plotgroups(
  groups,
  samdat,
  xmin = 0,
  xmax = 0,
  sitecol = "site",
  length = "length",
  mat10 = "mature"
)
}
\arguments{
\item{groups}{a list containing vectors of site numbers}

\item{samdat}{a data.frame of maturity data containing the required site(s)}

\item{xmin}{a generic x-axis minimum?, if 0 use local, default=0}

\item{xmax}{a generic x-axis maximum?, if 0 use local, default=0}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}
}
\value{
an invisible list containing the parameters and the fitmaturity 
    results. It also generates a plot.
}
\description{
plotgroups after using findgroups one can use plotgroups to 
    plot maturity curves to each group found. It uses fitgroups and outputs
    those results invisibly.
}
\examples{
print("wait on internal data sets")
#  groups=groups1; samdat=samb;length="length";mat10="mature";sitecol="site"
}
