% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{plotmaturity}
\alias{plotmaturity}
\title{plotmaturity generates a plot of fitted maturity data.}
\usage{
plotmaturity(model, label = "", col = 2, xmin = 0, xmax = 0, setpar = FALSE)
}
\arguments{
\item{model}{the output from the fitmaturity function}

\item{label}{the label for the y-axis, usually the site number or the
combination of site numbers making up a group}

\item{col}{colour of the predicted maturity curve, default=2 = red}

\item{xmin}{a generic x-axis minimum?, if 0 use local, default=0}

\item{xmax}{a generic x-axis maximum?, if 0 use local, default=0}

\item{setpar}{should parset be called, useful when only a single plot is
being made. default=FALSE}
}
\value{
nothing but it does generate a plot
}
\description{
plotmaturity generates a plot of the maturity curve and data 
   when provided with the model output from fitmaturity
}
\examples{
print("wait on data sets")
}
\seealso{
{
  \link{fitmaturity} 
}
}
