% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{plotsiteweights}
\alias{plotsiteweights}
\title{plotsiteweights plots each sites weight-at-length curve from an SAU}
\usage{
plotsiteweights(resblk, cex.lab = 0.75)
}
\arguments{
\item{resblk}{the output from fitsiteweights containing a list of fitexp
results}

\item{cex.lab}{The font size of the axis labels, default=0.75 = small}
}
\value{
Nothing but a plot is generated
}
\description{
plotsiteweights fits an exponential curve to the data from each 
    site within an sau and, optionally plots them, along with an analysis 
    and plot of the sau data combined.
}
\examples{
print("wait for data sets")
}
