% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{plottwogroups}
\alias{plottwogroups}
\title{plottwosites fits and plots two sites, including the combined by default}
\usage{
plottwogroups(
  grp1,
  grp2,
  samdat,
  length = "length",
  mat10 = "mature",
  sitecol = "site"
)
}
\arguments{
\item{grp1}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{grp2}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame containing the maturity data that includes the two 
sites}

\item{length}{column name of the length variable within the data.frame,
default='length'}

\item{mat10}{column name of the maturity variable containing mature=1, and
immature = 0 for all length observations. default='mature'}

\item{sitecol}{column name for the site number, default='combsite', ready for 
when sites are combined}
}
\value{
invisibly, a list containing the model fits for each site and the
    combined sites if combine = TRUE
}
\description{
plottwosites fits a maturity curve for two sites and, by default, 
    the combined data for both sites. It then plots them together on the same 
    graph. The out is, invisibly, a list of all results.
}
\examples{
print("wait on internal data sets")
# grp1=c(43,874,2);grp2=c(128); samdat=samW;length="length";mat10="mature";sitecol="site"
}
