% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{printgroup}
\alias{printgroup}
\title{printgroup a simple function to print the group list}
\usage{
printgroup(group)
}
\arguments{
\item{group}{a list of groups of sites from the maturity data. Usually
derived from the function 'findgroups'}
}
\value{
nothing but it does print out the group list more tidily
}
\description{
printgroup is a simple function to print a group of sites list
    that is produced by the function 'findgroups'. This is used instead of
    defining a new S3 method so one doesn't have to define a new class.
}
\examples{
args(printgroup)
agroup <- list("1"=c(1,2,3),"2"=c(4,5),"3"=c(6,7,8,9))
printgroup(agroup)
}
\seealso{
{
  \link{findgroups}
}
}
