% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othercurves.R
\name{sdpow}
\alias{sdpow}
\title{sdpow a power relationship used to describe the changing variance}
\usage{
sdpow(p, predDL)
}
\arguments{
\item{p}{the vector of parameters, what these are will depend on both grow
and sdfunc}

\item{predDL}{the predicted DeltaL derived from some growth function}
}
\value{
a vector of standard deviations the same length as predDL
}
\description{
sdpow a power relationship p[4]*predDL^p[5] used to describe
    how the standard deviation to be used when estimating normal likelihoods
    changes relative to the predicted growth increment
}
\examples{
 pars <- c(2.2,9.5,5,0.6,1)
 dat <- NULL  # wait on internal dataset
 # negLLG(p=pars,grow=invlog,sdfunc=sdpow)
}
