% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{selecttwo}
\alias{selecttwo}
\title{selecttwo selects data from a data.frame on two factors}
\usage{
selecttwo(samdat, value1, value2, field1 = "site", field2 = "season")
}
\arguments{
\item{samdat}{the data.frame being selected from.}

\item{value1}{a vector or values for the first field, possibly just one}

\item{value2}{a vector or values for the second field, possibly just one}

\item{field1}{the column name in the data.frame from which to select the
value1}

\item{field2}{the column name in the data.frame from which to select the
value2}
}
\value{
a data.frame subsetted from the input data.rame
}
\description{
selecttwo selects data from a data.frame on two factors/fields
    where the values for each field can be a vector of values. This can be 
    useful when there ar emultiple levels within multiple fields/factors and
    how best to combine them is unknown.
}
\examples{
\dontrun{
  site1 <- "George"
  seas1 <- c("Spring","Summer")
  dat1 <- selecttwo(samdat=sam,value1=site1,value2=c(seas1),
                    field1="site",field2="season")
}
}
