% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{siteprops}
\alias{siteprops}
\title{siteprops tabulates the properties of each site number in a data.frame}
\usage{
siteprops(samdat, sitecol = "site")
}
\arguments{
\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}
}
\value{
a data.frame of site properties
}
\description{
siteprops extracts the main properties of a vector of site
    numbers. These properties include: n, site, sitename, lat, long, block,
    subblock, date, and propM (proportion mature).
}
\examples{
print("wait on internal data sets")
# sits=bsites; samdat=samb
}
\seealso{
{
   \link{getloc}, \link{comparetwosites}
}
}
