% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{testgroup}
\alias{testgroup}
\title{testgroup literally compares sites using logistic regression}
\usage{
testgroup(grp, samdat, sitecol = "site", length = "length", mature = "mature")
}
\arguments{
\item{grp}{a vector of at least two site numbers to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mature}{column name for the maturity variable in samdat mature=1,
immature = 0}
}
\value{
a list of a summary of the logistic regression, and the model
}
\description{
testgroup using logistic regressions to compare whether 
    two or more sites are sufficiently similar for their data to be combined, 
    or not.
}
\examples{
args(testgroup)
print("wait on internal data sets")
# site1=grp;site2=grp2; sitecol="site";samdat=samb;length="length";mature="mature"
}
\seealso{
{
   \link{getloc}, \link{siteprops}
}
}
