% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{fitmaturity}
\alias{fitmaturity}
\title{fitmaturity fits a maturity ~ length logistic regression to maturity}
\usage{
fitmaturity(
  sams,
  length = "length",
  mature = "mature",
  lower = 50,
  upper = 160
)
}
\arguments{
\item{sams}{a data.frame of maturity data. This requires that any data 
selection of say sites and/or seasons must be done outside the function.}

\item{length}{column name in sams for the length variable}

\item{mature}{column name for the maturity variable in sams mature=1,
immature = 0}

\item{lower}{the smallest size class for inclusion in the predicted plot 
data, the default = 50}

\item{upper}{the largest size class for inclusion in the predicted plot 
data, the default = 160}
}
\value{
an invisible list containing the regression results, the model,
    the parameters, and the plotdata, which contains both predm and 95 CI.
}
\description{
fitmaturity uses data of the form maturity = 1 immaturity = 0,
    and length and fits a glm using binomial errors (a logistic regression).
    It outputs the regression results, the parameters, a, b, Lm50, and IQ, as
    well as the data for plotting as 'lens', 'propm', and 'predm'
}
\examples{
args(fitmaturity)
data(tasab)
# sams=samdat;length="length";mature="mature";lower=50;upper=160
}
