% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{getsiteparams}
\alias{getsiteparams}
\title{getsiteparams get all the individual site parameters in a data.frame}
\usage{
getsiteparams(samdat, sitecol = "site", length = "length", mat10 = "mature")
}
\arguments{
\item{samdat}{a data.frame of maturity data containing sites}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}

\item{length}{column name in samdat for the length variable}

\item{mat10}{column name for the maturity variable in samdat mature=1,
immature = 0}
}
\value{
a list containing a matrix of the site parameters and a list of all
    the model fits used (called allfit)
}
\description{
getsiteparams estimates the site parmaeters and all the model
    fits for all sites within a data.frame. This sets up the system ready
    for site comparisons
}
\examples{
print("wait on intenral data sets")
}
