% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{groupprops}
\alias{groupprops}
\title{groupprops tabulates the properties of each site number}
\usage{
groupprops(grp, samdat, groupname = "grp", sitecol = "site")
}
\arguments{
\item{grp}{a vector of at least one site number to be found in the 
data.frame samdat. The default column name = 'site'.}

\item{samdat}{a data.frame of maturity data with site numbers}

\item{groupname}{a name for the group, default='grp'}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}
}
\value{
a list of the sites, the block(s), n, propI and propM
}
\description{
groupprops extracts the main properties of two vectors of site
    numbers. These properties include: n, site, sitename, lat, long, block,
    subblock, date, and propM (proportion mature). It uses siteprops
}
\examples{
print("wait on internal data sets")
# grp=c(780,783); samdat=samw; sitecol="site"; groupname="blk13"
}
\seealso{
{
   \link{getloc}, \link{findgroups}, \link{fitgroups}
}
}
