% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturitycalc.r
\name{plottwogroups}
\alias{plottwogroups}
\title{plottwosites fits and plots two sites, including the combined by default}
\usage{
plottwogroups(
  sam1,
  sam2,
  length = "length",
  mature = "mature",
  sitecol = "site"
)
}
\arguments{
\item{sam1}{a data.frame containing the maturity data for group 1}

\item{sam2}{a data.frame containing the maturity data for group 2}

\item{length}{column name of the length variable within the data.frame,
default='length'}

\item{mature}{column name of the maturity variable containing mature=1, and
immature = 0 for all length observations. default='mature'}

\item{sitecol}{column name of the variable for the site numbers. The default
column name = 'site'}
}
\value{
invisibly, a list containing the model fits for each group of sites, 
    and the model for all sites combined
}
\description{
plottwosites fits a maturity curve for two sites and, by default, 
    the combined data for both sites. It then plots them together on the same 
    graph. The out is, invisibly, a list of all results.
}
\examples{
args(plottwogroups)
print("wait on internal data sets")
# grp1=c(43,874,2);grp2=c(128); samdat=samW;length="length";mature="mature";sitecol="site"
}
