% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{projtozero}
\alias{projtozero}
\title{projtozero extends the predicted inverse logistic to zero}
\usage{
projtozero(ans)
}
\arguments{
\item{ans}{the output from fitIL}
}
\value{
nothing but does add a line to a plot from plotmodelIL
}
\description{
projtozero extends the predicted inverse logistic derived from
    fitIL to zero.
}
